moved Config package "config" and renamed "load" to "model"
